+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Kayah Li numeral system
 - Gunjala Gondi numeral system
 - Masaram Gondi numeral system
+- Mro numeral system
 ## [0.9] - 2025-10-10
 ### Added
 - Dives Akuru numeral system